Rerported ads and block add

+++ aclassdeal/src/main/java/com/anaadih/aclassdeal/Repository/ProductRepository.java
@@ -10,7 +10,7 @@ import com.anaadih.aclassdeal.Model.CategoryModel;
 import com.anaadih.aclassdeal.Model.ProductModel;
 
 @Repository
-public interface ProductRepository extends PagingAndSortingRepository<ProductModel,Long>{
+public interface ProductRepository extends PagingAndSortingRepository<ProductModel,Integer>{
 
 	List<ProductModel> findByStatus(String status);
